chore: jsx 페이지 세팅
